Remove redundant check for python support
The check for python in line 18 is incomplete as it checks only for python2 whereas the plugin clearly supports python3. For those of us who have python3 support but not python2, your plugin doesn't load. Besides, python support is being checked for, with a helpful error message in the next code block.